Extract handler to separate class, and change header set ordering
