+++ pbj-core/pbj-grpc-helidon/src/main/java/com/hedera/pbj/grpc/helidon/PbjProtocolSelector.java
@@ -146,57 +146,4 @@ class PbjProtocolSelector implements Http2SubProtocolSelector {
                         route,
                         deadlineDetector));
     }
-
-    /** A handler for the case where the path is not found. */
-    private static final class RouteNotFoundHandler
-            implements Http2SubProtocolSelector.SubProtocolHandler {
-        private final Http2StreamWriter streamWriter;
-        private final int streamId;
-        private Http2StreamState currentStreamState;
-
-        RouteNotFoundHandler(
-                final Http2StreamWriter streamWriter,
-                final int streamId,
-                final Http2StreamState currentStreamState) {
-            this.streamWriter = streamWriter;
-            this.streamId = streamId;
-            this.currentStreamState = currentStreamState;
-        }
-
-        @Override
-        public void init() {
-            WritableHeaders<?> writable = WritableHeaders.create();
-            writable.set(Http2Headers.STATUS_NAME, Status.OK_200.code());
-            writable.set(GrpcHeaders.NOT_FOUND);
-            Http2Headers http2Headers = Http2Headers.create(writable);
-            streamWriter.writeHeaders(
-                    http2Headers,
-                    streamId,
-                    Http2Flag.HeaderFlags.create(
-                            Http2Flag.END_OF_HEADERS | Http2Flag.END_OF_STREAM),
-                    FlowControl.Outbound.NOOP);
-            currentStreamState = Http2StreamState.HALF_CLOSED_LOCAL;
-        }
-
-        @Override
-        @NonNull
-        public Http2StreamState streamState() {
-            return currentStreamState;
-        }
-
-        @Override
-        public void rstStream(Http2RstStream rstStream) {
-            // No-op
-        }
-
-        @Override
-        public void windowUpdate(@NonNull final Http2WindowUpdate update) {
-            // No-op
-        }
-
-        @Override
-        public void data(@NonNull final Http2FrameHeader header, @NonNull final BufferData data) {
-            // No-op
-        }
-    }
 }
